jllooper: new looper.
this loopeoper does not depend on any metronome. Simple interface: one button to start and stop recording, one button to start and stop playback. To do: add automatic loop length based on bpm and number of bars (independent of any external metronome). array size is now 10 seconds, must be changed to a multiple of that number pure data likes, which I do not remember now.